feature(pagination): edit pageInfo totalItems -> size

+++ src/main/java/kr/hhplus/be/server/api/config/PageInfo.java
@@ -1,5 +1,5 @@
 package kr.hhplus.be.server.api.config;
 
-public record PageInfo(int currentPage, int size, int totalItems) {
+public record PageInfo(int currentPage, int size, long totalItems) {
 
 }
